# users/api.yaml
openapi: 3.0.3
info:
  title: Users
  description: Users microservice responsible for storing account data.
  version: 1.0.0
servers:
  - url: 'http://localhost:8011'
tags:
  - name: Users
    description: User management
paths:
  /users:
    post:
      tags:
        - Users
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        200:
          description: User created successfully

        400:
          description: Invalid request content

    get:
      tags:
        - Users
      parameters:
        - in: query
          name: id
          style: form
          explode: false
          required: false
          schema:
            type: array
            items:
              type: string
              format: uuid

      responses:
        200:
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

      summary: Return a list of users
  /users/get/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Users
      summary: Fetch specific user profile
      responses:
        200:
          description: Show user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        404:
          description: User not found
    delete:
      tags:
        - Users
      summary: Delete a user and all their data
      responses:
        200:
          description: User deleted successfully
        404:
          description: User not found
components:
  schemas:
    CreateUser:
      required:
        - username
        - email
      properties:
        email:
          type: string
          format: email
        username:
          type: string
    User:
      properties:
        username:
          type: string
          example: adam.kowalski
        email:
          type: string
          format: email
        name:
          type: string
          example: Adam Kowalski
        id:
          type: string
          format: uuid